refactor new quote and docket menus
