+++ frontend/src/components/NewQuote/NewQuote.scss
@@ -1,0 +1,103 @@
+.quotespage-container {
+  display: flex;
+}
+
+.new-quote-menu-container {
+  width: 20vw;
+  height: max-content;
+  background-color: var(--dark-grey);
+  margin: 1vh 0.5vw;
+  border-radius: 15px;
+  height: 100%;
+}
+
+.new-quote-title {
+  text-align: center;
+  h1 {
+    padding-top: 0.25rem;
+    font-size: 2rem;
+  }
+
+  font-weight: bold;
+  color: var(--light-accent);
+  margin: 0.75vh 0;
+  user-select: none;
+}
+
+.new-quote-fields {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: space-evenly;
+}
+
+.new-quote-field {
+  width: 80%;
+  height: 2rem;
+  border: 1px solid #000;
+  border-radius: 15px;
+  padding: 0 1vw;
+  margin: 0.5vh 0;
+  font-size: 16px;
+}
+
+.new-quote-submit {
+  width: 80%;
+  height: 5vh;
+  border: 1px solid #000;
+  border-radius: 15px;
+  padding: 0 1vw;
+  margin: 0.5vh 0 1vh 0;
+  font-size: 20px;
+  background-color: #728ef4;
+  color: var(--white);
+  cursor: pointer;
+  &:hover {
+    background-color: var(--light-accent);
+    color: var(--dark-grey);
+    transition: background-color 200ms;
+  }
+}
+
+.quotes-list-container {
+  width: 100%;
+  margin: 1vh 0.5vw;
+  border-radius: 15px;
+  overflow: auto;
+}
+
+.new-quote-customer-select {
+  width: 80%;
+  height: 2rem;
+  border: 1px solid #000;
+  border-radius: 15px;
+  padding: 0 1vw;
+  margin: 0.5vh 0;
+  font-size: 16px;
+  background-color: var(--white);
+  color: var(--dark-grey);
+  cursor: pointer;
+}
+
+.new-quote-customer-select__menu {
+  background-color: none;
+  color: var(--dark-grey);
+}
+.new-quote-customer-select__menu-list {
+  border: 1px solid #000;
+
+  border-radius: 15px;
+  padding: 0.75vh;
+  width: 80%;
+  background-color: field;
+  color: var(--dark-grey);
+}
+
+.new-quote-customer-select__option {
+  border-radius: 15px;
+  padding: 0.5vh 0.5vw;
+  &:hover,
+  &:active {
+    background-color: lightblue;
+  }
+}
